@@ -14,11 +14,11 @@
     "clsx": "^2.1.1",
     "immer": "^10.1.1",
     "lucide-react": "^0.525.0",
-    "next": "15.3.5",
+    "next": "^16.0.10",
     "node-appwrite": "^17.0.0",
     "postcss": "^8.5.6",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
+    "react": "^19.2.3",
+    "react-dom": "^19.2.3",
     "tailwind-merge": "^3.3.1",
     "zustand": "^5.0.6"
   },
